mcp-fixer-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
openai==1.47.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
mcp==1.0.0
